Updating splat

+++ include/types.h
@@ -7,7 +7,7 @@ typedef int int32_t;
 typedef long long int64_t;
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
-typedef unsigned int uint32_t;
+typedef unsigned long uint32_t;
 typedef unsigned long long uint64_t;
 
 typedef char s8;
@@ -16,7 +16,8 @@ typedef int s32;
 typedef long long s64;
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned int u32;
+// long is 4 bytes :S
+typedef unsigned long u32;
 typedef unsigned long long u64;
 typedef float f32;
 typedef double f64;
